Extract components, use jsx instead of createElement
